+++ jet_tracking/gui/widgets/jetImageWidget.py
@@ -1,8 +1,8 @@
 from PyQt5.QtWidgets import QFrame, QGraphicsPixmapItem
 from PyQt5.QtGui import QPixmap, QImage
 from gui.widgets.jetImageWidgetUi import Image_Ui
-import qimage2ndarray as qi
-import cv2
+#import qimage2ndarray as qi
+#import cv2
 import logging
 import numpy as np
